[REFACTOR/#170] pilgrimage 패키지 분리
