+++ src/main/java/com/favoriteplace/app/community/domain/GuestBook.java
@@ -3,7 +3,7 @@ package com.favoriteplace.app.community.domain;
 import com.favoriteplace.app.domain.Image;
 import com.favoriteplace.app.member.domain.Member;
 import com.favoriteplace.app.domain.common.BaseTimeEntity;
-import com.favoriteplace.app.domain.travel.Pilgrimage;
+import com.favoriteplace.app.pilgrimage.domain.Pilgrimage;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
