+++ src/main/java/com/favoriteplace/app/community/repository/GuestBookRepository.java
@@ -2,7 +2,7 @@ package com.favoriteplace.app.community.repository;
 
 import com.favoriteplace.app.member.domain.Member;
 import com.favoriteplace.app.community.domain.GuestBook;
-import com.favoriteplace.app.domain.travel.Pilgrimage;
+import com.favoriteplace.app.pilgrimage.domain.Pilgrimage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
