+++ src/main/java/com/favoriteplace/app/community/service/GuestBookCommandService.java
@@ -5,17 +5,17 @@ import com.favoriteplace.app.member.domain.Member;
 import com.favoriteplace.app.community.domain.GuestBook;
 import com.favoriteplace.app.community.domain.HashTag;
 import com.favoriteplace.app.member.domain.enums.PointType;
-import com.favoriteplace.app.domain.travel.Pilgrimage;
-import com.favoriteplace.app.domain.travel.VisitedPilgrimage;
+import com.favoriteplace.app.pilgrimage.domain.Pilgrimage;
+import com.favoriteplace.app.pilgrimage.domain.VisitedPilgrimage;
 import com.favoriteplace.app.community.controller.dto.guestbook.GuestBookModifyRequestDto;
 import com.favoriteplace.app.community.controller.dto.PostResponseDto;
 import com.favoriteplace.app.community.repository.GuestBookRepository;
 import com.favoriteplace.app.community.repository.HashtagRepository;
 import com.favoriteplace.app.repository.ImageRepository;
 import com.favoriteplace.app.community.repository.LikedPostRepository;
-import com.favoriteplace.app.repository.PilgrimageRepository;
+import com.favoriteplace.app.pilgrimage.repository.PilgrimageRepository;
 import com.favoriteplace.app.item.repository.PointHistoryRepository;
-import com.favoriteplace.app.repository.VisitedPilgrimageRepository;
+import com.favoriteplace.app.pilgrimage.repository.VisitedPilgrimageRepository;
 import com.favoriteplace.global.exception.ErrorCode;
 import com.favoriteplace.global.exception.RestApiException;
 import com.favoriteplace.global.gcpImage.UploadImage;
