+++ src/main/java/com/favoriteplace/app/community/service/GuestBookQueryService.java
@@ -3,13 +3,13 @@ package com.favoriteplace.app.community.service;
 import com.favoriteplace.app.community.converter.GuestBookConverter;
 import com.favoriteplace.app.member.domain.Member;
 import com.favoriteplace.app.community.domain.GuestBook;
-import com.favoriteplace.app.domain.travel.Pilgrimage;
+import com.favoriteplace.app.pilgrimage.domain.Pilgrimage;
 import com.favoriteplace.app.community.controller.dto.guestbook.GuestBookResponseDto;
 import com.favoriteplace.app.community.controller.dto.TrendingPostResponseDto;
 import com.favoriteplace.app.community.repository.GuestBookImplRepository;
 import com.favoriteplace.app.community.repository.LikedPostRepository;
-import com.favoriteplace.app.repository.PilgrimageRepository;
-import com.favoriteplace.app.repository.VisitedPilgrimageRepository;
+import com.favoriteplace.app.pilgrimage.repository.PilgrimageRepository;
+import com.favoriteplace.app.pilgrimage.repository.VisitedPilgrimageRepository;
 import com.favoriteplace.app.community.service.searchStrategy.SearchGuestBookByContent;
 import com.favoriteplace.app.community.service.searchStrategy.SearchGuestBookByNickname;
 import com.favoriteplace.app.community.service.searchStrategy.SearchGuestBookByTitle;
