+++ src/main/java/com/favoriteplace/app/pilgrimage/controller/PilgrimageApiController.java
@@ -1,11 +1,11 @@
-package com.favoriteplace.app.controller;
+package com.favoriteplace.app.pilgrimage.controller;
 
 import com.favoriteplace.app.member.domain.Member;
 import com.favoriteplace.app.dto.CommonResponseDto;
-import com.favoriteplace.app.dto.travel.PilgrimageResponseDto;
+import com.favoriteplace.app.pilgrimage.controller.dto.PilgrimageResponseDto;
 import com.favoriteplace.app.dto.travel.RallyResponseDto;
-import com.favoriteplace.app.service.PilgrimageCommandService;
-import com.favoriteplace.app.service.PilgrimageQueryService;
+import com.favoriteplace.app.pilgrimage.service.PilgrimageCommandService;
+import com.favoriteplace.app.pilgrimage.service.PilgrimageQueryService;
 import com.favoriteplace.global.exception.ErrorCode;
 import com.favoriteplace.global.exception.RestApiException;
 import com.favoriteplace.global.util.SecurityUtil;
