+++ src/main/java/com/favoriteplace/app/pilgrimage/controller/PilgrimageSocketController.java
@@ -1,10 +1,10 @@
-package com.favoriteplace.app.controller;
+package com.favoriteplace.app.pilgrimage.controller;
 
 import com.favoriteplace.app.member.domain.Member;
 import com.favoriteplace.app.dto.CommonResponseDto;
-import com.favoriteplace.app.dto.travel.PilgrimageCertifyRequestDto;
-import com.favoriteplace.app.dto.travel.PilgrimageSocketDto;
-import com.favoriteplace.app.service.PilgrimageCommandService;
+import com.favoriteplace.app.pilgrimage.controller.dto.PilgrimageCertifyRequestDto;
+import com.favoriteplace.app.pilgrimage.controller.dto.PilgrimageSocketDto;
+import com.favoriteplace.app.pilgrimage.service.PilgrimageCommandService;
 import com.favoriteplace.global.exception.ErrorCode;
 import com.favoriteplace.global.exception.RestApiException;
 import com.favoriteplace.global.security.CustomUserDetails;
