+++ src/main/java/com/favoriteplace/app/pilgrimage/controller/dto/PilgrimageCertifyRequestDto.java
@@ -1,4 +1,4 @@
-package com.favoriteplace.app.dto.travel;
+package com.favoriteplace.app.pilgrimage.controller.dto;
 
 public record PilgrimageCertifyRequestDto(Double longitude, Double latitude) {
 }
