+++ src/main/java/com/favoriteplace/app/pilgrimage/converter/PilgrimageConverter.java
@@ -1,11 +1,11 @@
-package com.favoriteplace.app.converter;
+package com.favoriteplace.app.pilgrimage.converter;
 
 import com.favoriteplace.app.domain.Image;
 import com.favoriteplace.app.community.domain.GuestBook;
 import com.favoriteplace.app.domain.travel.Address;
-import com.favoriteplace.app.domain.travel.Pilgrimage;
+import com.favoriteplace.app.pilgrimage.domain.Pilgrimage;
 import com.favoriteplace.app.domain.travel.Rally;
-import com.favoriteplace.app.dto.travel.PilgrimageResponseDto;
+import com.favoriteplace.app.pilgrimage.controller.dto.PilgrimageResponseDto;
 
 import java.util.ArrayList;
 import java.util.List;
