+++ src/main/java/com/favoriteplace/app/pilgrimage/repository/PilgrimageRepository.java
@@ -1,7 +1,7 @@
-package com.favoriteplace.app.repository;
+package com.favoriteplace.app.pilgrimage.repository;
 
 import com.favoriteplace.app.domain.travel.Address;
-import com.favoriteplace.app.domain.travel.Pilgrimage;
+import com.favoriteplace.app.pilgrimage.domain.Pilgrimage;
 import com.favoriteplace.app.domain.travel.Rally;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
