+++ src/main/java/com/favoriteplace/app/pilgrimage/repository/VisitedPilgrimageRepository.java
@@ -1,9 +1,9 @@
-package com.favoriteplace.app.repository;
+package com.favoriteplace.app.pilgrimage.repository;
 
 import com.favoriteplace.app.member.domain.Member;
-import com.favoriteplace.app.domain.travel.Pilgrimage;
+import com.favoriteplace.app.pilgrimage.domain.Pilgrimage;
 import com.favoriteplace.app.domain.travel.Rally;
-import com.favoriteplace.app.domain.travel.VisitedPilgrimage;
+import com.favoriteplace.app.pilgrimage.domain.VisitedPilgrimage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
