+++ src/main/java/com/favoriteplace/app/pilgrimage/service/PilgrimageCommandService.java
@@ -1,4 +1,4 @@
-package com.favoriteplace.app.service;
+package com.favoriteplace.app.pilgrimage.service;
 
 import com.favoriteplace.app.converter.CommonConverter;
 import com.favoriteplace.app.item.converter.PointHistoryConverter;
@@ -9,20 +9,20 @@ import com.favoriteplace.app.domain.enums.RallyVersion;
 import com.favoriteplace.app.item.domain.AcquiredItem;
 import com.favoriteplace.app.domain.travel.CompleteRally;
 import com.favoriteplace.app.domain.travel.LikedRally;
-import com.favoriteplace.app.domain.travel.Pilgrimage;
+import com.favoriteplace.app.pilgrimage.domain.Pilgrimage;
 import com.favoriteplace.app.domain.travel.Rally;
-import com.favoriteplace.app.domain.travel.VisitedPilgrimage;
+import com.favoriteplace.app.pilgrimage.domain.VisitedPilgrimage;
 import com.favoriteplace.app.dto.CommonResponseDto;
-import com.favoriteplace.app.dto.travel.PilgrimageCertifyRequestDto;
-import com.favoriteplace.app.dto.travel.PilgrimageSocketDto;
+import com.favoriteplace.app.pilgrimage.controller.dto.PilgrimageCertifyRequestDto;
+import com.favoriteplace.app.pilgrimage.controller.dto.PilgrimageSocketDto;
 import com.favoriteplace.app.item.repository.AcquiredItemRepository;
 import com.favoriteplace.app.repository.CompleteRallyRepository;
 import com.favoriteplace.app.community.repository.GuestBookRepository;
 import com.favoriteplace.app.repository.LikedRallyRepository;
-import com.favoriteplace.app.repository.PilgrimageRepository;
+import com.favoriteplace.app.pilgrimage.repository.PilgrimageRepository;
 import com.favoriteplace.app.item.repository.PointHistoryRepository;
 import com.favoriteplace.app.repository.RallyRepository;
-import com.favoriteplace.app.repository.VisitedPilgrimageRepository;
+import com.favoriteplace.app.pilgrimage.repository.VisitedPilgrimageRepository;
 import com.favoriteplace.app.notification.service.FCMNotificationService;
 import com.favoriteplace.global.exception.ErrorCode;
 import com.favoriteplace.global.exception.RestApiException;
